fix: rename some vercel-breaking modules
