+++ src/components/modules/LandingPageModule/sections/Faq.tsx
@@ -1,7 +1,7 @@
 import { Accordion } from 'flowbite-react'
-import { faqs } from './FaqConstant'
+import { FAQ_DATA } from './FaqConstant'
 
-export const FaqModule: React.FC = () => {
+export const FAQ: React.FC = () => {
   return (
     <>
       <div className="relative lg:h-screen flex flex-col items-center justify-center lg:px-44 px-5 py-5">
@@ -9,7 +9,7 @@ export const FaqModule: React.FC = () => {
           Butuh Bantuan? Baca FAQs
         </h1>
         <Accordion className="w-full border-transparent focus:border-transparent focus:ring-0 ">
-          {faqs.map((item, index) => (
+          {FAQ_DATA.map((item, index) => (
             <Accordion.Panel key={index}>
               <Accordion.Title className="text-black focus:bg-mintGreen border-transparent focus:border-transparent focus:ring-0 rounded-t-xl">
                 <div>
