+++ src/components/modules/LandingPageModule/sections/WalkingText.tsx
